fix : 조회된 건수에 따른 페이징 숫자 오류

+++ src/main/java/com/ebstudy/ebsoft/mvc/domain/PageInfo.java
@@ -11,11 +11,11 @@ public class PageInfo {
     private Boolean hasNextButton;
     private Integer lastPageNumber;
     private Integer firstPageNumber;
-    private Integer rightPageNumber;
+    private Double rightPageNumber;
     private Integer leftPageNumber;
 
     @Builder
-    public PageInfo(Integer currentPageNumber, Boolean hasPrevButton, Boolean hasNextButton, Integer lastPageNumber, Integer firstPageNumber, Integer rightPageNumber, Integer leftPageNumber) {
+    public PageInfo(Integer currentPageNumber, Boolean hasPrevButton, Boolean hasNextButton, Integer lastPageNumber, Integer firstPageNumber, Double rightPageNumber, Integer leftPageNumber) {
         this.currentPageNumber = currentPageNumber;
         this.hasPrevButton = hasPrevButton;
         this.hasNextButton = hasNextButton;
